fix some png images being discarded

+++ inc/image.php
@@ -363,9 +363,10 @@ class ImageConvert extends ImageBase {
 					$this->height,
 					escapeshellarg($this->temp)))) || !file_exists($this->temp)) {
 
-					if (strpos($error, "known incorrect sRGB profile") === false) {
+					if (strpos($error, "known incorrect sRGB profile") === false &&
+                                            strpos($error, "iCCP: Not recognizing known sRGB profile that has been edited") === false) {
 						$this->destroy();
-						error('Failed to resize image!', null, array('convert_error' => $error));
+						error(_('Failed to resize image!')." "._('Details: ').nl2br(htmlspecialchars($error)), null, array('convert_error' => $error));
 					}
 					if (!file_exists($this->temp)) {
 						$this->destroy();
